+++ lib/markdown/mdPluginSeparateForBlock.d.ts
@@ -1,2 +1,0 @@
-declare const _default: (md: MarkdownIt) => void;
-export default _default;